+++ src/index.ts
@@ -1518,6 +1518,7 @@ export {
   RawMongoCollection,
   MeteorMongoCollection,
   ensureIndexes,
+  ensureNormalizedNameUniqueIndex,
   scopeToFilter,
   mergeFilters,
   factFilterToMongo,